not important

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,13 +1,14 @@
 #include <stdio.h>
 #include "function_pointers.h"
 /**
+ * array_iterator - function to do actions on elements
  * @array: the array
  * @size: number of elemnts
  * @action: the function_pointer
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-long unsigned int i;
+size_t i;
 if (array == NULL || size == 0 || action == NULL)
 return;
 for (i = 0; i < size; i++)
